feat(verify-conditions): verifies conditions for pushing docker images
The verifyConditions plugin verifies that environment variables for authentication via username and
password are set. It uses a registry server provided via config or environment variable (preferred)
or defaults to docker hub if none is given.  It also verifies that the credentials are correct by
logging in to the given registry.